feat(13-15): implement erasure-scrub reclaim tick (Task 1)
- packages/shared-schemas/src/queues.ts: erasureScrubReclaimTickJobSchema +
  ERASURE_SCRUB_RECLAIM_TICK_SCHEMA_VERSION, mirroring sendReconcilerTickJobSchema
- apps/worker/src/queues/erasure-scrub-reclaim.worker.ts: scheduled tick
  (5min interval, upsertJobScheduler shape) that enumerates workspaces via
  withCrossWorkspaceScan, finds pending/scrubbing erasure_records past a
  15-minute lease under each workspace's own withTenant scope, and
  re-enqueues onto ERASURE_SCRUB_QUEUE through the shared
  buildErasureScrubJobPayload/buildErasureScrubJobId derivation plan 13-10's
  deleteContact already uses -- so a reclaim of an already-queued record
  collides on jobId instead of duplicating the scrub
- complete/failed records are excluded; a per-workspace enqueue failure is
  caught and logged so it does not abort the remaining workspaces
- test fixture bugfix during RED->GREEN: make_interval(mins => ...) requires
  an integer argument, not float8 -- explicit ::int casts in the seed helper

+++ apps/worker/src/queues/__tests__/erasure-scrub-reclaim.test.ts
@@ -84,8 +84,8 @@ describe("erasure-scrub-reclaim.worker.ts (CMP-04, D-04, plan 13-15)", () => {
           `INSERT INTO erasure_records (workspace_id, contact_id, anonymized_at, status, requested_at, scrub_started_at, scrub_error)
            VALUES (
              $1, $2, now(), $3,
-             now() - make_interval(mins => $4),
-             CASE WHEN $5::float8 IS NULL THEN NULL ELSE now() - make_interval(mins => $5) END,
+             now() - make_interval(mins => $4::int),
+             CASE WHEN $5::int IS NULL THEN NULL ELSE now() - make_interval(mins => $5::int) END,
              $6
            )
            RETURNING id`,
